fix: finish remote data source problems

+++ repository/src/main/java/com/cairosquad/repository/movie/data_source/remote/dto/AuthorDetailsDto.kt
@@ -13,7 +13,7 @@ data class AuthorDetailsDto(
     @SerialName("avatar_path")
     val avatarPath: String? = null,
     @SerialName("rating")
-    val rating: Int? = null
+    val rating: Double? = null
 ) {
 
 }
